Feat: Adicionando TopBar e pages
