+++ AquariumQualityControl/App.xaml
@@ -4,6 +4,6 @@
              xmlns:local="clr-namespace:AquariumQualityControl"
              StartupUri="MainWindow.xaml">
     <Application.Resources>
-         
+
     </Application.Resources>
 </Application>
