+++ AquariumQualityControl/App.xaml.cs
@@ -2,9 +2,6 @@
 
 namespace AquariumQualityControl;
 
-/// <summary>
-/// Interaction logic for App.xaml
-/// </summary>
 public partial class App : Application
 {
 }
