--- AquariumQualityControl/Components/MenuLayout.xaml
@@ -1,24 +1,0 @@
-﻿<UserControl x:Class="AquariumQualityControl.Components.MenuLayout"
-             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml">
-    
-    <Grid x:Name="menuGrid"
-          Width="150"
-          Background="#0288D1"
-          HorizontalAlignment="Left">
-        
-        <StackPanel>
-            
-            <Button Click="ToggleMenu" Background="{x:Null}" Height="30" Width="30" BorderBrush="#0288D1">
-                <Image Source="/Images/toggle-menu.png" Stretch="Uniform" Height="20" Width="20"/>
-            </Button>
-            
-            <Button Content="Item 1"/>
-            <Button Content="Item 2"/>
-            <Button Content="Item 3"/>
-
-        </StackPanel>
-        
-    </Grid>
-    
-</UserControl>
