+++ AquariumQualityControl/MainWindow.xaml.cs
@@ -1,4 +1,5 @@
 ﻿using System.Windows;
+using System.Windows.Input;
 
 namespace AquariumQualityControl;
 
@@ -8,4 +9,47 @@ public partial class MainWindow : Window
     {
         InitializeComponent();
     }
+
+    #region TopBar Methods
+
+    private void CloseClick(object sender, RoutedEventArgs e)
+    {
+        Close();
+    }
+
+    private void TopBar_MouseLeftButtonDown(object sender, MouseButtonEventArgs e)
+    {
+        if (e.ChangedButton is MouseButton.Left)
+            DragMove();
+    }
+
+    private void TopBar_MouseMove(object sender, MouseEventArgs e)
+    {
+        if (e.LeftButton is MouseButtonState.Pressed)
+        {
+            Point currentPosition = e.GetPosition(null);
+            Left = currentPosition.X - (Width / 2);
+            Top = currentPosition.Y - 15;
+        }
+    }
+
+    #endregion
+
+    #region Load pages Methods
+
+    private void Load_HomePage(object sender, RoutedEventArgs e)
+    {
+        MainFrame.Navigate(new Uri(
+            "/Pages/HomePage.xaml",
+            UriKind.Relative));
+    }
+
+    private void Load_AmmoniaPage(object sender, RoutedEventArgs e)
+    {
+        MainFrame.Navigate(new Uri(
+            "/Pages/AmmoniaPage.xaml",
+            UriKind.Relative));
+    }
+
+    #endregion
 }
