+++ AquariumQualityControl/Pages/AmmoniaPage.xaml.cs
@@ -1,0 +1,11 @@
+﻿using System.Windows.Controls;
+
+namespace AquariumQualityControl.Pages;
+
+public partial class AmmoniaPage : Page
+{
+    public AmmoniaPage()
+    {
+        InitializeComponent();
+    }
+}
