+++ AquariumQualityControl/Pages/HomePage.xaml
@@ -1,0 +1,14 @@
+﻿<Page x:Class="AquariumQualityControl.Pages.HomePage"
+      xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+      xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+      xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
+      xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+      xmlns:local="clr-namespace:AquariumQualityControl.Pages"
+      mc:Ignorable="d" 
+      d:DesignHeight="450" d:DesignWidth="800"
+      Title="HomePage">
+
+    <Grid Background="red">
+        <TextBlock HorizontalAlignment="Center" VerticalAlignment="Center" FontSize="40">Page teste 1</TextBlock>
+    </Grid>
+</Page>
