+++ AquariumQualityControl/Pages/HomePage.xaml.cs
@@ -1,0 +1,11 @@
+﻿using System.Windows.Controls;
+
+namespace AquariumQualityControl.Pages;
+
+public partial class HomePage : Page
+{
+    public HomePage()
+    {
+        InitializeComponent();
+    }
+}
